ceph-disk: tolerate partprobe racing with udev
After we change the final partition type, sgdisk may or may not trigger a
udev event, depending on how well udev is behaving (it varies between
distros, it seems).  The old code would often settle and wait for udev to
activate the device, and then partprobe would uselessly fail because it
was already mounted.  Instead, run partprobde unconditionally, ignore the
result, and skip the settle entirely.

Fixes: #4948
Backport: cuttlefish
Signed-off-by: Sage Weil <sage@inktank.com>